Moved Banner component to /components. Refactored the code.
