+++ app/jobs/dependencies_check_job.rb
@@ -2,7 +2,6 @@ class DependenciesCheckJob < ActiveJob::Base
   queue_as :default
 
   def perform
-    Rails.cache.write(:dependencies_report, nil)
     request = HTTPI::Request.new
     request.url = ENV["MONITOR_URL"]
     http = HTTPI.get(request, :httpclient)
