+++ app/services/dependencies_check.rb
@@ -5,17 +5,14 @@ class DependenciesCheck
     def find_degraded_dependencies
       begin
         report = JSON.parse Rails.cache.read(:dependencies_report)
-        dependencies_outage = []
-        report.values.each do |hash|
-          if hash["up_rate_5"].to_i < 51
-            dependencies_outage << hash["name"]
-            Rails.logger.error "Dependencies outage: #{hash["name"]}"
-          end
+        report.values.reduce([]) do |result, element|
+          Rails.logger.error "Dependencies outage: #{element["name"]}"
+          result << element["name"] if element["up_rate_5"].to_i < 50
+          result
         end
       rescue
          Rails.logger.info "Invalid response from #{ENV["MONITOR_URL"]}"
       end
-      dependencies_outage
     end
 
     def outage_present?
