+++ client/app/constants/Constants.js
@@ -1,0 +1,1 @@
+export const POLLING_INTERVAL = 30000;
