+++ config/environments/development.rb
@@ -45,9 +45,6 @@ Rails.application.configure do
   #
   ENV["CASEFLOW_FEEDBACK_URL"] = "https://dsva-appeals-feedback-demo-1748368704.us-gov-west-1.elb.amazonaws.com/"
 
-  # Used for performace degradation banner
-  ENV["MONITOR_URL"] = "http://monitor.cf.uat.ds.va.gov/sample"
-
   ENV["METRICS_USERNAME"] ||= "caseflow"
   ENV["METRICS_PASSWORD"] ||= "caseflow"
   ENV["SIDEKIQ_USERNAME"] ||= "caseflow"
