+++ config/environments/production.rb
@@ -82,9 +82,6 @@ Rails.application.configure do
   config.google_analytics_account = "UA-74789258-1"
   config.google_analytics_host = "//www.google-analytics.com/analytics.js"
 
-  # Used for performace degradation banner
-  ENV["MONITOR_URL"] = "http://monitor.cf.ds.va.gov/sample"
-
   config.google_analytics_domains = [
     "appeals#{ENV["COOKIE_DOMAIN"]}",
     "certification#{ENV["COOKIE_DOMAIN"]}"
